Add v2.0 support for CDXComponent object
