+++ cmd/component_test.go
@@ -232,21 +232,3 @@ func TestComponentListCdx16EmptyDefnsDecls(t *testing.T) {
 	innerTestComponentList(t, ti, COMPONENT_TEST_DEFAULT_FLAGS)
 }
 
-// -------------------------------------------
-// CDX v2.0 variants
-// -------------------------------------------
-
-// TestComponentListCdx20ValidComponentPartiesCsv verifies that "component list" for a CycloneDX
-// v2.0 BOM correctly reads supplier, manufacturer, and publisher from the component's v2.0
-// "parties" array (not the legacy metadata.supplier / metadata.manufacturer direct fields).
-func TestComponentListCdx20ValidComponentPartiesCsv(t *testing.T) {
-	utils.GlobalFlags.ValidateFlags.SchemaVariant = SCHEMA_VARIANT_DEVELOPMENT
-	defer func() { utils.GlobalFlags.ValidateFlags.SchemaVariant = SCHEMA_VARIANT_NONE }()
-
-	ti := NewComponentTestInfoBasic(TEST_COMPONENT_LIST_CDX_2_0_VALID_COMPONENT, FORMAT_CSV, nil)
-	ti.ResultExpectedLineCount = 3 // title + 1 data row + EOF LF
-	ti.ResultLineContainsValuesAtLineNum = 1
-	// All three party roles must be sourced from parties[], not legacy direct fields
-	ti.ResultLineContainsValues = []string{"Acme Supplier Ltd", "Acme Manufacturing Inc", "Acme Corp"}
-	innerTestComponentList(t, ti, COMPONENT_TEST_DEFAULT_FLAGS)
-}
